+++ src/include/kernel/process.h
@@ -14,15 +14,15 @@
 #define PROC_USER_HEAP  0x80000  // 512 Kib (initial size, can be expanded)
 
 typedef struct process {
-    uint32_t pid; // ID
-    uint32_t esp; // current kernel stack position
-    uint32_t ss; // current kernel stack segment
-    uint32_t kern_stack; // where kernel stack is allocated
-    uint32_t user_stack; // where user stack is allocated
-    uint32_t cr3; // current page directory
-    uint32_t time_to_run; // time allocated for process
-    uint32_t ready; // ready or not?
-    uint32_t prior; // priority
+    size_t pid; // ID
+    size_t esp; // current kernel stack position
+    size_t ss; // current kernel stack segment
+    size_t kern_stack; // where kernel stack is allocated
+    size_t user_stack; // where user stack is allocated
+    size_t cr3; // current page directory
+    size_t time_to_run; // time allocated for process
+    size_t ready; // ready or not?
+    size_t prior; // priority
     char name[32]; // name of process
     
     struct file **opened_files;
